Close.

+++ src/main/java/net/fexcraft/mod/fvtm/util/Axe.java
@@ -39,7 +39,7 @@ public class Axe {
         return new Vec3f(mat.m00, mat.m10, mat.m20);
     }
 
-	public Vec3f getRelativeVector(float x, int y, int z){
+	public Vec3f getRelativeVector(float x, float y, float z){
         Matrix4f mat = new Matrix4f(); mat.m00 = x; mat.m10 = y; mat.m20 = z;
         Matrix4f.rotate((roll  * 3.14159265F / 180f), new Vector3f(1F, 0F, 0F), mat, mat);
         Matrix4f.rotate((pitch * 3.14159265F / 180f), new Vector3f(0F, 0F, 1F), mat, mat);
